refactor(feature): refactor history-analytics feature - history-analytics feature 의 뷰들을 컴포넌트로 분리합니다.
